Move list-related commands into their own module
